requirements.txt
aiohttp==3.9.3
aiosignal==1.3.2
altair==5.5.0
asyncio==3.4.3
attrs==25.1.0
aws-lambda-powertools==1.25.5
aws-xray-sdk==2.14.0
blinker==1.9.0
boto3==1.37.0
botocore==1.37.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
et_xmlfile==2.0.0
fastjsonschema==2.21.1
fonttools==4.56.0
frozenlist==1.5.0
fyers-apiv2==2.0.9
fyers_apiv3==3.1.5
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
holidays==0.70
idna==2.10
Jinja2==3.1.6
jmespath==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mplfinance==0.12.10b0
multidict==6.1.0
mysql-connector-python==9.2.0
narwhals==1.33.0
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.3.0
protobuf==5.29.4
pyarrow==19.0.1
pycparser==2.22
pydeck==0.9.1
PyMySQL==1.1.1
pyotp==2.9.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
s3transfer==0.11.2
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.38
streamlit==1.44.1
tenacity==9.1.2
toml==0.10.2
tornado==6.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
watchdog==6.0.0
websocket-client==1.2.1
websockets==8.1
wrapt==1.17.2
XlsxWriter==3.2.2
yarl==1.18.3
